Refactor: 어노테이션 추가 및 메서드 폴더 이동
