+++ src/main/java/cloudcomputinginha/demo/service/member/MemberCommandServiceImpl.java
@@ -39,14 +39,7 @@ public class MemberCommandServiceImpl implements MemberCommandService {
     }
 
     @Override
-    public MemberInfoResponseDTO getBasicInfo(Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-
-        return MemberConverter.toMemberInfoResponseDTO(member);
-    }
-
-    @Override
+    @Transactional
     public void deleteMember(Long memberId) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
