+++ src/main/java/cloudcomputinginha/demo/service/member/MemberQueryService.java
@@ -1,0 +1,7 @@
+package cloudcomputinginha.demo.service.member;
+
+import cloudcomputinginha.demo.web.dto.MemberInfoResponseDTO;
+
+public interface MemberQueryService {
+    MemberInfoResponseDTO getBasicInfo(Long memberId);
+}
