+++ src/main/java/cloudcomputinginha/demo/web/controller/MemberRestController.java
@@ -2,6 +2,7 @@ package cloudcomputinginha.demo.web.controller;
 
 import cloudcomputinginha.demo.apiPayload.ApiResponse;
 import cloudcomputinginha.demo.service.member.MemberCommandService;
+import cloudcomputinginha.demo.service.member.MemberQueryService;
 import cloudcomputinginha.demo.web.dto.MemberInfoRequestDTO;
 import cloudcomputinginha.demo.web.dto.MemberInfoResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,6 +20,7 @@ import org.springframework.web.bind.annotation.*;
 @Validated
 public class MemberRestController {
     private final MemberCommandService memberCommandService;
+    private final MemberQueryService memberQueryService;
 
     @PatchMapping("")
     @Operation(summary = "사용자 기본 정보 등록 API", description = "사용자의 기본 정보(전화번호, 직무 분야, 자기소개)를 등록합니다.")
@@ -37,7 +39,7 @@ public class MemberRestController {
     @GetMapping
     @Operation(summary = "사용자 프로필 조회 API", description = "사용자의 프로필을 조회합니다.")
     public ApiResponse<MemberInfoResponseDTO> getBasicInfo(@AuthenticationPrincipal Long memberId) {
-        MemberInfoResponseDTO result = memberCommandService.getBasicInfo(memberId);
+        MemberInfoResponseDTO result = memberQueryService.getBasicInfo(memberId);
         return ApiResponse.onSuccess(result);
     }
